Refactor itemper constants and modules
